requirements.txt
beautifulsoup4>=4.*
requests>=2.*
azure-cosmos>=4.*
azure-eventgrid~=4.9
